requirements.txt
apollo-client==0.8.2
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
dnspython==1.16.0
eventlet==0.25.1
google-auth==1.8.1
greenlet==0.4.15
idna==2.8
kubernetes==10.0.1
monotonic==1.5
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
websocket-client==0.56.0
